Move OpenStack compute driver to its own module.
